#82 Clean up TMDB shows and movies domain module
