+++ domain/tmdb-movies/src/main/java/io/filmtime/domain/tmdb/movies/GetSimilarMoviesUseCase.kt
@@ -1,10 +1,10 @@
-package io.filmtime.domain.tmdb.shows
+package io.filmtime.domain.tmdb.movies
 
 import io.filmtime.data.model.GeneralError
 import io.filmtime.data.model.Result
 import io.filmtime.data.model.VideoThumbnail
 
-interface GetSimilarUseCase {
+interface GetSimilarMoviesUseCase {
 
   suspend operator fun invoke(movieId: Int): Result<List<VideoThumbnail>, GeneralError>
 }
