+++ domain/tmdb-shows/src/main/java/io/filmtime/domain/tmdb/shows/GetSimilarShowsUseCase.kt
@@ -1,10 +1,10 @@
-package io.filmtime.domain.tmdb.movies
+package io.filmtime.domain.tmdb.shows
 
 import io.filmtime.data.model.GeneralError
 import io.filmtime.data.model.Result
 import io.filmtime.data.model.VideoThumbnail
 
-interface GetSimilarUseCase {
+interface GetSimilarShowsUseCase {
 
   suspend operator fun invoke(movieId: Int): Result<List<VideoThumbnail>, GeneralError>
 }
