+++ domain/tmdb-shows/src/main/java/io/filmtime/domain/tmdb/shows/impl/GetShowCreditsUseCaseImpl.kt
@@ -1,12 +1,13 @@
-package io.filmtime.domain.tmdb.shows
+package io.filmtime.domain.tmdb.shows.impl
 
 import io.filmtime.data.model.GeneralError
 import io.filmtime.data.model.Person
 import io.filmtime.data.model.Result
 import io.filmtime.data.tmdb.shows.TmdbShowsRepository
+import io.filmtime.domain.tmdb.shows.GetShowCreditsUseCase
 import javax.inject.Inject
 
-class GetShowCreditsUseCaseImpl @Inject constructor(
+internal class GetShowCreditsUseCaseImpl @Inject constructor(
   private val tmdbMovieRepository: TmdbShowsRepository,
 ) : GetShowCreditsUseCase {
