+++ domain/tmdb-shows/src/main/java/io/filmtime/domain/tmdb/shows/impl/GetShowDetailsUseCaseImpl.kt
@@ -1,9 +1,10 @@
-package io.filmtime.domain.tmdb.shows
+package io.filmtime.domain.tmdb.shows.impl
 
 import io.filmtime.data.model.GeneralError
 import io.filmtime.data.model.Result
 import io.filmtime.data.model.VideoDetail
 import io.filmtime.data.tmdb.shows.TmdbShowsRepository
+import io.filmtime.domain.tmdb.shows.GetShowDetailsUseCase
 import javax.inject.Inject
 
 internal class GetShowDetailsUseCaseImpl @Inject constructor(
