+++ domain/tmdb-shows/src/main/java/io/filmtime/domain/tmdb/shows/impl/GetShowsListUseCaseImpl.kt
@@ -1,4 +1,4 @@
-package io.filmtime.domain.tmdb.shows
+package io.filmtime.domain.tmdb.shows.impl
 
 import io.filmtime.data.model.GeneralError
 import io.filmtime.data.model.Result
@@ -12,6 +12,7 @@ import io.filmtime.data.model.VideoListType.Trending
 import io.filmtime.data.model.VideoListType.Upcoming
 import io.filmtime.data.model.VideoThumbnail
 import io.filmtime.data.tmdb.shows.TmdbShowsRepository
+import io.filmtime.domain.tmdb.shows.GetShowsListUseCase
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 import javax.inject.Inject
