+++ domain/tmdb-shows/src/main/java/io/filmtime/domain/tmdb/shows/impl/GetSimilarShowsUseCaseImpl.kt
@@ -1,13 +1,16 @@
-package io.filmtime.domain.tmdb.shows
+package io.filmtime.domain.tmdb.shows.impl
 
 import io.filmtime.data.model.GeneralError
 import io.filmtime.data.model.Result
 import io.filmtime.data.model.VideoThumbnail
 import io.filmtime.data.tmdb.shows.TmdbShowsRepository
+import io.filmtime.domain.tmdb.shows.GetSimilarShowsUseCase
 import javax.inject.Inject
 
-class GetSimilarUseCaseImpl @Inject constructor(private val tmdbMovieRepository: TmdbShowsRepository) :
-  GetSimilarUseCase {
+internal class GetSimilarShowsUseCaseImpl @Inject constructor(
+  private val tmdbMovieRepository: TmdbShowsRepository,
+) : GetSimilarShowsUseCase {
+
   override suspend fun invoke(movieId: Int): Result<List<VideoThumbnail>, GeneralError> =
     tmdbMovieRepository.getSimilar(movieId)
 }
